(#385) Fix  wrong position for Drag Adorner on High DPI screens
