+++ src/GongSolutions.WPF.DragDrop/Utilities/WindowStyleHelper.cs
@@ -1,93 +1,13 @@
 ﻿using System;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-// Mouse move hook inspired by Fluent Drag&Drop https://github.com/punker76/FluentDragDrop
 namespace GongSolutions.Wpf.DragDrop.Utilities
 {
-    internal static class NativeMethods
+    internal static class WindowStyleHelper
     {
-        private delegate IntPtr LowLevelMouseProc(int nCode, IntPtr wParam, IntPtr lParam);
-
-        private const int WH_MOUSE_LL = 14;
-
-        private static IntPtr _hookID = IntPtr.Zero;
-        private static Action<System.Windows.Point> _mouseMoveHandler;
-
-        // wee need to keep the variable to prevent the GarbageCollector to remove the HookCallback
-        // https://social.msdn.microsoft.com/Forums/vstudio/en-US/68fdc3dc-8d77-48c4-875c-5312baa56aee/how-to-fix-callbackoncollecteddelegate-exception?forum=netfxbcl
-        private static LowLevelMouseProc _proc = HookCallback;
-
-        internal static IntPtr HookMouseMove(Action<System.Windows.Point> mouseMoveHandler)
-        {
-            _mouseMoveHandler = mouseMoveHandler;
-
-            using var process = Process.GetCurrentProcess();
-            using var module = process.MainModule;
-            return SetWindowsHookEx(WH_MOUSE_LL, _proc, GetModuleHandle(module.ModuleName), 0);
-        }
-
-        internal static void RemoveHook(IntPtr hookId)
-        {
-            UnhookWindowsHookEx(hookId);
-        }
-
-        private static IntPtr HookCallback(int nCode, IntPtr wParam, IntPtr lParam)
-        {
-            if (nCode >= 0 && MouseMessages.WM_MOUSEMOVE == (MouseMessages)wParam)
-            {
-                MSLLHOOKSTRUCT hookStruct = (MSLLHOOKSTRUCT)Marshal.PtrToStructure(lParam, typeof(MSLLHOOKSTRUCT));
-                //Debug.WriteLine(hookStruct.pt.x + ", " + hookStruct.pt.y);
-                var mousePosScreen = new System.Windows.Point(hookStruct.pt.x, hookStruct.pt.y);
-                _mouseMoveHandler?.Invoke(mousePosScreen);
-            }
-
-            return CallNextHookEx(_hookID, nCode, wParam, lParam);
-        }
-
-        private enum MouseMessages
-        {
-            WM_LBUTTONDOWN = 0x0201,
-            WM_LBUTTONUP = 0x0202,
-            WM_MOUSEMOVE = 0x0200,
-            WM_MOUSEWHEEL = 0x020A,
-            WM_RBUTTONDOWN = 0x0204,
-            WM_RBUTTONUP = 0x0205
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        private struct POINT
-        {
-            public int x;
-            public int y;
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        private struct MSLLHOOKSTRUCT
-        {
-            public POINT pt;
-            public uint mouseData;
-            public uint flags;
-            public uint time;
-            public IntPtr dwExtraInfo;
-        }
-
-        [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-        private static extern IntPtr SetWindowsHookEx(int idHook, LowLevelMouseProc lpfn, IntPtr hMod, uint dwThreadId);
-
-        [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        private static extern bool UnhookWindowsHookEx(IntPtr hhk);
-
-        [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-        private static extern IntPtr CallNextHookEx(IntPtr hhk, int nCode, IntPtr wParam, IntPtr lParam);
-
-        [DllImport("kernel32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-        private static extern IntPtr GetModuleHandle(string lpModuleName);
-
         [CLSCompliant(false)]
         internal static WS_EX GetWindowStyleEx(IntPtr hWnd)
         {
@@ -103,7 +23,7 @@ namespace GongSolutions.Wpf.DragDrop.Utilities
         // This is aliased as a macro in 32bit Windows.
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode")]
         [MethodImpl(MethodImplOptions.NoInlining)]
-        internal static IntPtr GetWindowLongPtr(IntPtr hwnd, GWL nIndex)
+        private static IntPtr GetWindowLongPtr(IntPtr hwnd, GWL nIndex)
         {
             var ret = 8 == IntPtr.Size
                 ? GetWindowLongPtr64(hwnd, nIndex)
@@ -130,7 +50,7 @@ namespace GongSolutions.Wpf.DragDrop.Utilities
         // This is aliased as a macro in 32bit Windows.
         [SuppressMessage("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode")]
         [MethodImpl(MethodImplOptions.NoInlining)]
-        internal static IntPtr SetWindowLongPtr(IntPtr hwnd, GWL nIndex, IntPtr dwNewLong)
+        private static IntPtr SetWindowLongPtr(IntPtr hwnd, GWL nIndex, IntPtr dwNewLong)
         {
             if (8 == IntPtr.Size)
             {
